SERVER - Add Basic HTTP Server Setup
Using civetweb which comes with Zephyr OS to provide a basic HTTP server
and the ability to toggle the relay state using a URL.  This requires
some libc extensions which have been added in the lib directory.  Extra
configuration has been added to the kernel configuration file
(prj.conf).

An IP address and netmask have been hard-coded in prj.conf but this will
be replaced and made configurable in a future commit.