SHELL - Enable UART/Telnet Shell Feature
Enabling the Zephyr OS shell feature with its telnet backend so that we
have a CLI over both the USB UART and via traditional telnet.  Actions
that can be performed via the web GUI can now be made to work via the
CLI which may suit better for scripting, etc.