TERMINAL SERVER
Some basic terminal server support (with all debug left in):
This uses a configuration somewhat similar to ser2net - it's an array
that maps UARTs to TCP ports and configuration settings so the same UART
can be used at different speeds etc when accessed via different ports.
See the dut_serial_cfg array in main.c.

Also moved the relay GPIO to D7 so we can use the UART on D0/D1.