TFTP SERVER - Initial Implementation
Added code to provide a TFTP server on the DUTCHess.  This is planned
for use in rewriting the images in EEPROMs on the DUT via TFTP.  It may
also see use as a general TFTP server.  Currently only a test handler is
provided but the system is extendable with new handlers which decouple
the server from the storage media.